app/requirements.txt
coveralls==1.11.1
flask==2.1.1
google-cloud-secret-manager==1.0.0
google-cloud-bigquery==3.0.1
gunicorn==19.10.0
requests==2.24.0
sendgrid==6.5.0
